[Refactor] test파일 분리(test파일이 길어지면서 가독성 저하
