+++ test_shell/basic_test.cpp
@@ -1,0 +1,296 @@
+#pragma once
+#ifdef _DEBUG
+#include "test_case.h"
+
+TEST_F(TestShellFixture, ExitCMD) {
+    string cmdInput = EXIT_COMMAND_NAME;
+    testing::internal::CaptureStdout();
+    testShell.runCommand(cmdInput);
+    std::string output = testing::internal::GetCapturedStdout();
+
+    EXPECT_NE(output.find("PROGRAM EXIT"), std::string::npos);
+}
+
+TEST_F(TestShellFixture, InvalidCMD) {
+    string cmdInput = TEST_SCRIPT_2_FULL_COMMAND_NAME;
+    testing::internal::CaptureStdout();
+    testShell.runCommand(cmdInput);
+    std::string output = testing::internal::GetCapturedStdout();
+
+    EXPECT_EQ(output.find(INVALID_COMMAND_MSG), std::string::npos);
+
+    cmdInput = "XXXX";
+    testing::internal::CaptureStdout();
+    testShell.runCommand(cmdInput);
+    output = testing::internal::GetCapturedStdout();
+
+    EXPECT_NE(output.find(INVALID_COMMAND_MSG), std::string::npos);
+}
+
+TEST_F(TestShellFixture, ReadPass) {
+    EXPECT_CALL(mockSSDAdapter, read(10))
+        .Times(1)
+        .WillRepeatedly(Return("0xAAAABBBB"));
+
+    EXPECT_EQ("[Read] LBA 10 : 0xAAAABBBB", testShell.read(10));
+}
+
+TEST_F(TestShellFixture, ReadFailWrongLBA) {
+    EXPECT_CALL(mockSSDAdapter, read(100))
+        .Times(1)
+        .WillOnce(Return("ERROR"));
+
+    EXPECT_EQ("[Read] ERROR", testShell.read(100));
+}
+
+TEST_F(TestShellFixture, FullReadPass) {
+    internal::CaptureStdout();
+    EXPECT_CALL(mockSSDAdapter, read(_))
+        .Times(100)
+        .WillRepeatedly(Return("0x00ABCDEF"));
+
+    string expected;
+    for (int LBA = 0; LBA < 100; LBA++) {
+        expected += "[Read] LBA " + std::to_string(LBA) + " : " + "0x00ABCDEF\n";
+    }
+
+    testShell.fullRead();
+    string output = internal::GetCapturedStdout();
+
+    EXPECT_EQ(expected, output);
+}
+
+TEST_F(TestShellFixture, FullReadFail) {
+    internal::CaptureStdout();
+    EXPECT_CALL(mockSSDAdapter, read(_))
+        .Times(100)
+        .WillOnce(Return("0x00ABCDEF"))
+        .WillRepeatedly(Return("ERROR"));
+
+    string expected = "[Read] LBA 0 : 0x00ABCDEF\n";
+    for (int LBA = 1; LBA < 100; LBA++) {
+        expected += "[Read] ERROR\n";
+    }
+
+    testShell.fullRead();
+    string output = internal::GetCapturedStdout();
+
+    EXPECT_EQ(expected, output);
+}
+
+TEST_F(TestShellFixture, WritePass) {
+    EXPECT_CALL(mockSSDAdapter, write(5, "0xAAAABBBB"))
+        .WillOnce(Return(""));
+
+    std::string result = testShell.write(5, "0xAAAABBBB");
+    EXPECT_EQ(result, "[Write] Done");
+}
+
+TEST_F(TestShellFixture, WriteFail) {
+    EXPECT_CALL(mockSSDAdapter, write(5, "0xAAAAABBBBB"))
+        .WillOnce(Return("ERROR"));
+
+    std::string result = testShell.write(5, "0xAAAAABBBBB");
+    EXPECT_EQ(result, "[Write] ERROR");
+}
+
+TEST_F(TestShellFixture, FullWritePass) {
+    internal::CaptureStdout();
+    for (int i = 0; i < 100; ++i) {
+        EXPECT_CALL(mockSSDAdapter, write(i, "0xAAAABBBB"))
+            .Times(1)
+            .WillRepeatedly(Return(""));
+    }
+
+    testShell.fullWrite("0xAAAABBBB");
+    std::string output = internal::GetCapturedStdout();
+
+    EXPECT_NE(output.find("[fullWrite] Done"), std::string::npos);
+}
+
+TEST_F(TestShellFixture, FullWriteFail) {
+    internal::CaptureStdout();
+    EXPECT_CALL(mockSSDAdapter, write(_, _))
+        .WillOnce(Return(""))
+        .WillOnce(Return(""))
+        .WillOnce(Return(""))
+        .WillOnce(Return("[Write] ERROR"));
+
+    testShell.fullWrite("0xAAAABBBB");
+    std::string output = internal::GetCapturedStdout();
+
+    EXPECT_NE(output.find("[fullWrite] Failed at LBA 3"), std::string::npos);
+}
+
+TEST_F(TestShellFixture, EraseWithSize) {
+
+    EXPECT_CALL(mockSSDAdapter, erase(_, _))
+        .Times(1)
+        .WillRepeatedly(Return(""));
+
+    testing::internal::CaptureStdout();
+    testShell.runCommand("erase 0 10");
+    std::string output = testing::internal::GetCapturedStdout();
+
+    EXPECT_NE(output.find("[Erase] Done"), std::string::npos);
+}
+
+TEST_F(TestShellFixture, EraseWithInvalidSize) {
+
+    testing::internal::CaptureStdout();
+    testShell.runCommand("erase 0 101");
+    std::string output = testing::internal::GetCapturedStdout();
+
+    EXPECT_EQ(output.find("[Erase] Done"), std::string::npos);
+    EXPECT_NE(output.find("[Erase] Error"), std::string::npos);
+}
+
+TEST_F(TestShellFixture, EraseWithSize_CheckEraseCount) {
+
+    EXPECT_CALL(mockSSDAdapter, erase(_, _))
+        .Times(1)
+        .WillRepeatedly(Return(""));
+
+    testing::internal::CaptureStdout();
+    testShell.runCommand("erase 0 10");
+    std::string output = testing::internal::GetCapturedStdout();
+
+    //
+
+    EXPECT_CALL(mockSSDAdapter, erase(_, _))
+        .Times(2)
+        .WillRepeatedly(Return(""));
+
+    testing::internal::CaptureStdout();
+    testShell.runCommand("erase 0 11");
+    output = testing::internal::GetCapturedStdout();
+
+    //
+
+    EXPECT_CALL(mockSSDAdapter, erase(_, _))
+        .Times(5)
+        .WillRepeatedly(Return(""));
+
+    testing::internal::CaptureStdout();
+    testShell.runCommand("erase 1 50");
+    output = testing::internal::GetCapturedStdout();
+
+    //
+
+    EXPECT_CALL(mockSSDAdapter, erase(_, _))
+        .Times(6)
+        .WillRepeatedly(Return(""));
+
+    testing::internal::CaptureStdout();
+    testShell.runCommand("erase 2 53");
+    output = testing::internal::GetCapturedStdout();
+}
+
+TEST_F(TestShellFixture, EraseWithSize_EraseFailCase) {
+
+    EXPECT_CALL(mockSSDAdapter, erase(_, _))
+        .WillOnce(Return(""))
+        .WillOnce(Return(""))
+        .WillOnce(Return(""))
+        .WillOnce(Return(""))
+        .WillOnce(Return("ERROR"))
+        .WillRepeatedly(Return(""));
+
+
+    testing::internal::CaptureStdout();
+    testShell.runCommand("erase 0 100");
+    std::string output = testing::internal::GetCapturedStdout();
+
+    EXPECT_NE(output.find("[Erase] Error"), std::string::npos);
+}
+
+TEST_F(TestShellFixture, EraseRange_BaseTest) {
+
+    EXPECT_CALL(mockSSDAdapter, erase(_, _))
+        .Times(1)
+        .WillRepeatedly(Return(""));
+
+    testing::internal::CaptureStdout();
+    testShell.runCommand("erase_range 0 9");
+    std::string output = testing::internal::GetCapturedStdout();
+
+    EXPECT_NE(output.find("[Erase_Range] Done"), std::string::npos);
+}
+
+TEST_F(TestShellFixture, EraseRangeWithValidEndLBA_EdgeCase) {
+
+    testing::internal::CaptureStdout();
+    testShell.runCommand("erase_range 0 99");
+    std::string output = testing::internal::GetCapturedStdout();
+
+    EXPECT_NE(output.find("[Erase_Range] Done"), std::string::npos);
+    EXPECT_EQ(output.find("[Erase_Range] Error"), std::string::npos);
+}
+
+TEST_F(TestShellFixture, EraseRangeWithInvalidEndLBA) {
+
+    testing::internal::CaptureStdout();
+    testShell.runCommand("erase_range 0 100");
+    std::string output = testing::internal::GetCapturedStdout();
+
+    EXPECT_EQ(output.find("[Erase_Range] Done"), std::string::npos);
+    EXPECT_NE(output.find("[Erase_Range] Error"), std::string::npos);
+}
+
+TEST_F(TestShellFixture, EraseRange_CheckEraseCount) {
+
+    // expect calling erase one time
+    EXPECT_CALL(mockSSDAdapter, erase(_, _))
+        .Times(1)
+        .WillRepeatedly(Return(""));
+
+    testing::internal::CaptureStdout();
+    testShell.runCommand("erase_range 0 9");
+    std::string output = testing::internal::GetCapturedStdout();
+
+    // expect calling erase two times
+    EXPECT_CALL(mockSSDAdapter, erase(_, _))
+        .Times(2)
+        .WillRepeatedly(Return(""));
+
+    testing::internal::CaptureStdout();
+    testShell.runCommand("erase_range 0 10");
+    output = testing::internal::GetCapturedStdout();
+
+    // expect calling erase 5 times
+    EXPECT_CALL(mockSSDAdapter, erase(_, _))
+        .Times(5)
+        .WillRepeatedly(Return(""));
+
+    testing::internal::CaptureStdout();
+    testShell.runCommand("erase_range 2 51");
+    output = testing::internal::GetCapturedStdout();
+
+    // expect calling erase six times
+    EXPECT_CALL(mockSSDAdapter, erase(_, _))
+        .Times(6)
+        .WillRepeatedly(Return(""));
+
+    testing::internal::CaptureStdout();
+    testShell.runCommand("erase_range 2 53");
+    output = testing::internal::GetCapturedStdout();
+}
+
+TEST_F(TestShellFixture, EraseRange_EraseFailCase) {
+
+    EXPECT_CALL(mockSSDAdapter, erase(_, _))
+        .WillOnce(Return(""))
+        .WillOnce(Return(""))
+        .WillOnce(Return(""))
+        .WillOnce(Return(""))
+        .WillOnce(Return("ERROR"))
+        .WillRepeatedly(Return(""));
+
+
+    testing::internal::CaptureStdout();
+    testShell.runCommand("erase_range 0 55");
+    std::string output = testing::internal::GetCapturedStdout();
+
+    EXPECT_NE(output.find("[Erase_Range] Error"), std::string::npos);
+}
+#endif
